Projet 9 -/rag/app_simple/pyproject.toml
[project]
name = "advanced-interface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "chainlit>=2.8.3",
    "langchain-community>=0.3.30",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.8",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
]
